// lib/types/sanity.types.ts
export type TwoColumnRichText = {
  _type: 'twoColumnRichText';
  richText: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
};

export type TwoColumnList = {
  _type: 'twoColumnList';
  title?: string;
  items: Array<string>;
  showBackground?: boolean;
};

export type TwoCardSection = {
  _type: 'twoCardSection';
  cards: Array<
    {
      _key: string;
    } & ContentCard
  >;
};
// ...
export type PlainText = {
  _type: 'plainText';
  title: string;
  body?: string;
};
// ...
export type ImageRow = {
  _type: 'imageRow';
  image?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  title: string;
  content?: string;
  ctaText: string;
  ctaUrl: string;
  ctaColour?: 'navy' | 'teal' | 'orange' | 'white';
  ctaIcon?: 'phone' | 'email' | 'search';
  cta2text?: string;
  cta2url?: string;
  cta2colour?: 'navy' | 'teal' | 'orange' | 'white';
  cta2icon?: 'phone' | 'email' | 'search';
  dark?: boolean;
  overlayText: boolean;
};
// ...
export type GeneralInformation = {
  _id: string;
  _type: 'generalInformation';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  metadata?: Metadata;
  title: string;
  slug: Slug;
  parent?:
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'genericPage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'homepage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'featuredLocations';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'existingCustomers';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'landingPage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'featuredLocationsSub';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'findYourSpace';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'generalInformation';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'twoColumnEditorial';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'aboutUs';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'whyUs';
      };
  content?: Array<
    | ({
        _key: string;
      } & PlainText)
    | ({
        _key: string;
      } & RichText)
    | ({
        _key: string;
      } & CtaRichText)
    | ({
        _key: string;
      } & TwoColumnRichText)
    | ({
        _key: string;
      } & TwoCardSection)
    | ({
        _key: string;
      } & ThreeCardSection)
    | ({
        _key: string;
      } & FourCardSection)
    | ({
        _key: string;
      } & ImageRow)
    | ({
        _key: string;
      } & HorizontalContentCard)
    | ({
        _key: string;
      } & TwoColumnList)
    | ({
        _key: string;
      } & TwoColumnRichTextAndImage)
    | ({
        _key: string;
      } & TwoColumnRichTextAndForm)
    | ({
        _key: string;
      } & TwoSlotRichTextImage)
    | ({
        _key: string;
      } & FullWidthImage)
    | ({
        _key: string;
      } & VerticalTabs)
    | ({
        _key: string;
      } & PdfDownload)
    | ({
        _key: string;
      } & Spacer)
    | ({
        _key: string;
      } & RawHtml)
    | ({
        _key: string;
      } & VideoEmbed)
    | ({
        _key: string;
      } & InfoTable)
    | ({
        _key: string;
      } & GoogleReviews)
    | ({
        _key: string;
      } & CaseStudies)
  >;
};
// ...
export type LandingPage = {
  _id: string;
  _type: 'landingPage';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  metadata?: Metadata;
  title: string;
  slug: Slug;
  parent?:
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'genericPage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'homepage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'featuredLocations';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'existingCustomers';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'landingPage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'featuredLocationsSub';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'findYourSpace';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'generalInformation';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'twoColumnEditorial';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'aboutUs';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'whyUs';
      };
  pageIntroText: string;
  heroImage?: BrandfolderImage;
  content?: Array<
    | ({
        _key: string;
      } & PlainText)
    | ({
        _key: string;
      } & RichText)
    | ({
        _key: string;
      } & CtaRichText)
    | ({
        _key: string;
      } & TwoColumnRichText)
    | ({
        _key: string;
      } & TwoCardSection)
    | ({
        _key: string;
      } & ThreeCardSection)
    | ({
        _key: string;
      } & FourCardSection)
    | ({
        _key: string;
      } & ImageRow)
    | ({
        _key: string;
      } & HorizontalContentCard)
    | ({
        _key: string;
      } & TwoColumnList)
    | ({
        _key: string;
      } & TwoColumnRichTextAndImage)
    | ({
        _key: string;
      } & TwoColumnRichTextAndForm)
    | ({
        _key: string;
      } & TwoSlotRichTextImage)
    | ({
        _key: string;
      } & FullWidthImage)
    | ({
        _key: string;
      } & VerticalTabs)
    | ({
        _key: string;
      } & PdfDownload)
    | ({
        _key: string;
      } & Spacer)
    | ({
        _key: string;
      } & RawHtml)
    | ({
        _key: string;
      } & VideoEmbed)
    | ({
        _key: string;
      } & InfoTable)
    | ({
        _key: string;
      } & GoogleReviews)
    | ({
        _key: string;
      } & CaseStudies)
  >;
};

export type TwoColumnEditorial = {
  _id: string;
  _type: 'twoColumnEditorial';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  metadata?: Metadata;
  title: string;
  slug: Slug;
  parent?:
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'genericPage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'homepage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'featuredLocations';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'existingCustomers';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'landingPage';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'featuredLocationsSub';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'findYourSpace';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'generalInformation';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'twoColumnEditorial';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'aboutUs';
      }
    | {
        _ref: string;
        _type: 'reference';
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: 'whyUs';
      };
  heroImage?: BrandfolderImage;
  content?: Array<
    | ({
        _key: string;
      } & PlainText)
    | ({
        _key: string;
      } & RichText)
    | ({
        _key: string;
      } & CtaRichText)
    | ({
        _key: string;
      } & TwoColumnRichText)
    | ({
        _key: string;
      } & TwoCardSection)
    | ({
        _key: string;
      } & ThreeCardSection)
    | ({
        _key: string;
      } & FourCardSection)
    | ({
        _key: string;
      } & ImageRow)
    | ({
        _key: string;
      } & HorizontalContentCard)
    | ({
        _key: string;
      } & TwoColumnList)
    | ({
        _key: string;
      } & TwoColumnRichTextAndImage)
    | ({
        _key: string;
      } & TwoColumnRichTextAndForm)
    | ({
        _key: string;
      } & TwoSlotRichTextImage)
    | ({
        _key: string;
      } & FullWidthImage)
    | ({
        _key: string;
      } & VerticalTabs)
    | ({
        _key: string;
      } & PdfDownload)
    | ({
        _key: string;
      } & Spacer)
    | ({
        _key: string;
      } & RawHtml)
    | ({
        _key: string;
      } & VideoEmbed)
    | ({
        _key: string;
      } & InfoTable)
    | ({
        _key: string;
      } & GoogleReviews)
    | ({
        _key: string;
      } & CaseStudies)
  >;
  sideContent?: Array<
    | ({
        _key: string;
      } & ContentCard)
    | ({
        _key: string;
      } & RichText)
    | ({
        _key: string;
      } & LinkList)
  >;
};
// ...
export type ThreeCardSection = {
  _type: 'threeCardSection';
  cards: Array<
    {
      _key: string;
    } & ContentCard
  >;
};

export type FullWidthImage = {
  _type: 'fullWidthImage';
  image?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  fullHeight?: boolean;
};

export type FourCardSection = {
  _type: 'fourCardSection';
  cards: Array<
    {
      _key: string;
    } & ContentCard
  >;
};
// ...
export type CtaRichText = {
  _type: 'ctaRichText';
  pageIntroText: RichText;
  pageIntroCTATitle: string;
  pageIntroCTADescription: string;
  pageIntroCTALinkText: string;
  pageIntroCTALinkUrl: string;
};

export type RichText = {
  _type: 'richText';
  richText: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  centered?: boolean;
};
// ...
export type Slug = {
  _type: 'slug';
  current: string;
  source?: string;
};
// ...
export type Metadata = {
  _type: 'metadata';
  title?: string;
  description?: string;
  noIndex?: boolean;
  noFollow?: boolean;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;